feat(#41) :: Bookmark shared로 옮김
