+++ public/icons/shared/Bookmark.tsx
@@ -1,0 +1,52 @@
+import React from 'react';
+
+interface OwnProps {
+  width?: string;
+  height?: string;
+  stroke?: string;
+  fill?: string;
+  isChecked?: boolean;
+}
+
+const BookmarkIcon = ({ width, height, stroke, fill, isChecked }: OwnProps) => {
+  return (
+    <>
+      {!isChecked ? (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width={width ? width : '24'}
+          height={height ? height : '24'}
+          viewBox="0 0 19 18"
+          fill="none"
+        >
+          <path
+            fillRule="evenodd"
+            clipRule="evenodd"
+            d="M6.62339 1.94971H12.3765C12.7738 1.9497 13.1123 1.94969 13.3901 1.97239C13.682 1.99623 13.9659 2.04846 14.2373 2.18677C14.6466 2.39529 14.9793 2.72803 15.1878 3.13728C15.3262 3.40873 15.3784 3.69267 15.4022 3.98455C15.4249 4.26237 15.4249 4.60075 15.4249 4.99816V16.1247C15.4249 16.3685 15.2934 16.5934 15.0809 16.713C14.8685 16.8325 14.608 16.8282 14.3996 16.7017L9.49993 13.7269L4.60026 16.7017C4.39185 16.8282 4.1314 16.8325 3.91892 16.713C3.70644 16.5934 3.57495 16.3685 3.57495 16.1247V4.99816C3.57494 4.60077 3.57493 4.26236 3.59763 3.98455C3.62148 3.69267 3.6737 3.40873 3.81201 3.13728C4.02054 2.72803 4.35327 2.39529 4.76252 2.18677C5.03397 2.04846 5.31791 1.99623 5.60978 1.97239C5.88759 1.94969 6.22602 1.9497 6.62339 1.94971ZM6.34997 3.29976C5.73905 3.29976 5.59367 3.30809 5.49503 3.34014C5.24387 3.42175 5.04696 3.61866 4.96535 3.86982C4.9333 3.96846 4.92497 4.11384 4.92497 4.72476V14.9253L9.14965 12.3603C9.3649 12.2296 9.63501 12.2296 9.85027 12.3603L14.0749 14.9253V4.72476C14.0749 4.11384 14.0666 3.96846 14.0346 3.86982C13.953 3.61866 13.756 3.42175 13.5049 3.34014C13.4062 3.30809 13.2609 3.29976 12.6499 3.29976H6.34997Z"
+            fill={stroke ? stroke : 'white'}
+          />
+          <path
+            d="M5.49503 3.34014C5.59367 3.30809 5.73905 3.29976 6.34997 3.29976H12.6499C13.2609 3.29976 13.4062 3.30809 13.5049 3.34014C13.756 3.42175 13.953 3.61866 14.0346 3.86982C14.0666 3.96846 14.0749 4.11384 14.0749 4.72476V14.9253L9.85027 12.3603C9.63501 12.2296 9.3649 12.2296 9.14965 12.3603L4.92497 14.9253V4.72476C4.92497 4.11384 4.9333 3.96846 4.96535 3.86982C5.04696 3.61866 5.24387 3.42175 5.49503 3.34014Z"
+            fill={fill ? fill : '#C9C9C9'}
+            fillOpacity="0.2"
+          />
+        </svg>
+      ) : (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="24"
+          height="24"
+          viewBox="0 0 24 24"
+          fill="none"
+        >
+          <path
+            d="M8.06458 2C7.53473 1.99999 7.08353 1.99997 6.71311 2.03024C6.32395 2.06204 5.94535 2.13167 5.58342 2.31608C5.03776 2.59412 4.59411 3.03776 4.31608 3.58343C4.13167 3.94536 4.06203 4.32396 4.03024 4.71312C3.99997 5.08354 3.99999 5.53471 4 6.06457V20.9C4 21.2251 4.17531 21.5249 4.45862 21.6843C4.74193 21.8438 5.0892 21.838 5.36708 21.6693L11.9 17.7029L18.4329 21.6693C18.7107 21.838 19.058 21.8438 19.3413 21.6843C19.6246 21.5249 19.7999 21.2251 19.7999 20.9V6.0646C19.8 5.53473 19.8 5.08355 19.7697 4.71312C19.7379 4.32396 19.6683 3.94536 19.4839 3.58343C19.2058 3.03776 18.7622 2.59412 18.2165 2.31608C17.8546 2.13167 17.476 2.06204 17.0868 2.03024C16.7164 1.99997 16.2652 1.99999 15.7354 2H8.06458Z"
+            fill="#FF6A00"
+          />
+        </svg>
+      )}
+    </>
+  );
+};
+
+export default BookmarkIcon;
