+++ public/icons/shared/index.ts
@@ -5,3 +5,4 @@ export { default as ArrowIcon } from './Arrow';
 export { default as ShareIcon } from './Share';
 export { default as EditIcon } from './Edit';
 export { default as TrashIcon } from './Trash';
+export { default as BookmarkIcon } from './Bookmark';
